ML/requirements.txt
numpy
matplotlib
scikit-learn
scipy
pandas
tqdm
pytest
h5py
pytest-cov
pep8
jupyter
click
requests
lxml
bs4
sklearn
xgboost
mysql-connector
